Update release.ps1
Reapplied changes that for some reason did not get throuhgh

+++ release.ps1
@@ -38,9 +38,13 @@ if (Test-Path $outDir) {
 # Call msbuild clean
 & $msBuildPath /target:clean /p:Configuration=Release
 
-# Manually delete /bin and /obj directories
-Remove-Item -Path "$projDir/bin" -Recurse
-Remove-Item -Path "$projDir/obj" -Recurse
+# Manually delete /bin and /obj directories. Check that directories exist first.
+if (Test-Path "$projDir/bin") {
+    Remove-Item -Path "$projDir/bin" -Recurse
+}
+if (Test-Path "$projDir/obj") {
+    Remove-Item -Path "$projDir/obj" -Recurse
+}
 
 # Publish the application.
 Push-Location $projDir
